UI for serial port remote control
Serial port controllers are supported when:
 - 'config+=serial_remote_control' is specified in qmake invocation, or
 - '-DENABLE_SERIAL_REMOTE_CONTROL=1' is specified for cmake.

Note that QtSerialPort module must be available.

When the feature is enabled, two new actions are available in the menu and main
toolbar:
 - a toggle to open/close connection to a controller device
 - a button to open a configuration UI

The parameters of serial port have to be configured before a controller can be
used.
The configuration UI should be able to automatically detect all available
serial ports. Alternatively users can manually enter a point name if the correct
name was not detected.